menambah login session

+++ app/Controllers/Adbis.php
@@ -23,6 +23,9 @@ class Adbis extends BaseController
 
     public function index()
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
 
         $session = session();
         if ($session->get('role') == 3) {
@@ -103,6 +106,9 @@ class Adbis extends BaseController
 
     public function addFile()
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
         // dd($this->request->getFile('file'));
         $file = $this->request->getFile('file');
         if (!$file->isValid()) {
@@ -144,6 +150,9 @@ class Adbis extends BaseController
 
     public function deleteFile($id)
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
 
         $file = $this->filesModel->find($id);
         unlink('files/adbis/' . $file['store_file']);
@@ -173,6 +182,9 @@ class Adbis extends BaseController
 
     public function downloadFile($id)
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
         $file = $this->filesModel->find($id);
         // dd($file);
         return $this->response->download('files/adbis/' . $file['store_file'], null)->setFileName($file['file']);
