+++ app/Controllers/Akuntansi.php
@@ -23,6 +23,9 @@ class Akuntansi extends BaseController
 
     public function index()
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
 
         $session = session();
         if ($session->get('role') == 3) {
@@ -102,6 +105,9 @@ class Akuntansi extends BaseController
 
     public function addFile()
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
         // dd($this->request->getFile('file'));
         $file = $this->request->getFile('file');
         if (!$file->isValid()) {
@@ -144,6 +150,9 @@ class Akuntansi extends BaseController
     public function deleteFile($id)
     {
 
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
         $file = $this->filesModel->find($id);
         unlink('files/akuntansi/' . $file['store_file']);
 
@@ -172,6 +181,9 @@ class Akuntansi extends BaseController
 
     public function downloadFile($id)
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
         $file = $this->filesModel->find($id);
         // dd($file);
         return $this->response->download('files/akuntansi/' . $file['store_file'], null)->setFileName($file['file']);
