+++ app/Controllers/Sipil.php
@@ -23,6 +23,9 @@ class Sipil extends BaseController
 
     public function index()
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
 
         $session = session();
         if ($session->get('role') == 3) {
@@ -104,6 +107,9 @@ class Sipil extends BaseController
 
     public function addFile()
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
         // dd($this->request->getFile('file'));
         $file = $this->request->getFile('file');
         if (!$file->isValid()) {
@@ -145,6 +151,9 @@ class Sipil extends BaseController
 
     public function deleteFile($id)
     {
+        if (session('uname') == null) {
+            return redirect()->to('/auth');
+        }
 
         $file = $this->filesModel->find($id);
         unlink('files/sipil/' . $file['store_file']);
